refactor(agents): the personalization service keeps its methods, not its plumbing
chat-personalization-service.ts was 854 lines: nine service methods and
twenty-five helpers doing four different jobs behind them. The helpers now sit
with the job they belong to, and the service file is the nine methods.

Row reads and writes go to personalization-rows, which owns the repo. The
research lifecycle — claiming, staleness recovery, the daily guard, enqueueing
and the prefill lookup — goes to personalization-research. Role and website
normalisation goes to personalization-normalize, and the view assembly and
progress event to personalization-view.

The imports run one way: research reads rows, and the service reads all four.
IN_FLIGHT_STATUSES moved to rows, where both writers can reach it without the
two files importing each other.

Nothing is rewritten. Normalise away the imports and the new files match the
original line for line, in both directions.
