+++ packages/server/api/src/app/ee/agent/personalization/chat-personalization-service.ts
@@ -1,52 +1,30 @@
-import { ActivepiecesError, apId, ErrorCode, isNil, sanitizeObjectForPostgresql, spreadIfDefined, tryCatch } from '@activepieces/core-utils'
+import { apId, isNil, sanitizeObjectForPostgresql, spreadIfDefined, tryCatch } from '@activepieces/core-utils'
 import {
-    ApEdition,
     ChatPersonalization,
-    ChatPersonalizationProgressEvent,
     ChatPersonalizationScope,
     ChatPersonalizationStatus,
     ChatPersonalizationView,
     GetPersonalizationConfigRequest,
     GetPersonalizationPrefillConfigRequest,
-    LATEST_JOB_DATA_SCHEMA_VERSION,
     PersonalizationConfigResponse,
     PersonalizationPrefill,
     PersonalizationPrefillConfigResponse,
-    PersonalizationProfile,
-    PersonalizationScope,
-    PersonalizationUseCase,
-    PlatformRole,
     SavePersonalizationPrefillRequest,
     SavePersonalizationResultRequest,
     SendPersonalizationProgressRequest,
-    WebsocketClientEvent,
-    WorkerJobType,
 } from '@activepieces/shared'
 import { FastifyBaseLogger } from 'fastify'
 import { IsNull, Not } from 'typeorm'
-import { z } from 'zod'
-import { aiProviderService, ProviderScope } from '../../../ai/ai-provider-service'
 import { aiToolConfigService } from '../../../ai/ai-tool-config-service'
-import { repoFactory } from '../../../core/db/repo-factory'
-import { websocketService } from '../../../core/websockets.service'
 import { redisConnections } from '../../../database/redis-connections'
-import { system } from '../../../helper/system/system'
-import { AppSystemProp } from '../../../helper/system/system-props'
-import { assertCreditsAndAppSumoNotExceeded } from '../../../platform/billing-provider'
 import { platformService } from '../../../platform/platform.service'
 import { userService } from '../../../user/user-service'
-import { jobQueue, JobType } from '../../../workers/job-queue/job-queue'
 import { agentHelpers } from '../agent-helpers'
-import { ChatPersonalizationEntity } from './chat-personalization-entity'
+import { normalizeRoleTitle, normalizeWebsite } from './personalization-normalize'
+import { apolloApiKey, claimForResearch, enqueueResearchJob, guardsAllowResearch, PERSONALIZATION_PROVIDER_SCOPE, PREFILL_TTL_SECONDS, prefillKey, readPrefill, recoverIfStale, RESEARCH_STALENESS_TIMEOUT_MS, startPrefillLookup, validateResult } from './personalization-research'
+import { callerMayEditCompany, findRow, IN_FLIGHT_STATUSES, personalizationRepo, upsertFoundingUserRow, UpsertParams, writeCompanyRow, writeUserRow } from './personalization-rows'
+import { emitProgress, PersonalizationIdentity, toScopeEnum, toView } from './personalization-view'
 
-const personalizationRepo = repoFactory(ChatPersonalizationEntity)
-
-const PERSONALIZATION_PROVIDER_SCOPE: ProviderScope = { type: 'platform' }
-const RESEARCH_STALENESS_TIMEOUT_MS = 2 * 60 * 1_000
-const IN_FLIGHT_STATUSES = [ChatPersonalizationStatus.PENDING, ChatPersonalizationStatus.RESEARCHING]
-const RESEARCH_RUNS_PER_PLATFORM_PER_DAY = 5
-const RATE_LIMIT_TTL_SECONDS = 24 * 60 * 60
-const PREFILL_TTL_SECONDS = 7 * 24 * 60 * 60
 
 export const chatPersonalizationService = (log: FastifyBaseLogger) => ({
 
@@ -406,449 +384,3 @@ export const chatPersonalizationService = (log: FastifyBaseLogger) => ({
     },
 
 })
-
-function apolloApiKey(): string | null {
-    if (system.getEdition() !== ApEdition.CLOUD) {
-        return null
-    }
-    const key = system.get(AppSystemProp.APOLLO_API_KEY)
-    return isNil(key) || key.length === 0 ? null : key
-}
-
-async function findRow({ platformId, userId }: { platformId: string, userId: string | null }): Promise<ChatPersonalization | null> {
-    return personalizationRepo().findOneBy(
-        isNil(userId) ? { platformId, userId: IsNull() } : { platformId, userId },
-    )
-}
-
-async function upsertFoundingUserRow({ platformId, userId, researchToken, validated, log }: {
-    platformId: string
-    userId: string
-    researchToken: string | null
-    validated: ValidatedResult
-    log: FastifyBaseLogger
-}): Promise<void> {
-    if (isNil(researchToken)) {
-        return
-    }
-    const profile = validated.profile === null ? null : JSON.stringify(sanitizeObjectForPostgresql(validated.profile))
-    const useCases = validated.useCases === null ? null : JSON.stringify(sanitizeObjectForPostgresql(validated.useCases))
-    const seeded: { id: string }[] = await personalizationRepo().query(
-        `
-        INSERT INTO "chat_personalization" ("id", "created", "updated", "platformId", "userId", "domain", "companyText", "role", "status", "researchToken", "profile", "useCases")
-        SELECT $1, now(), now(), $2, $3, NULL, NULL, NULL, $4, NULL, $5::jsonb, $6::jsonb
-        WHERE EXISTS (
-            SELECT 1 FROM "chat_personalization"
-            WHERE "platformId" = $2 AND "userId" IS NULL AND "researchToken" = $7
-        )
-        ON CONFLICT ("platformId", "userId") WHERE "userId" IS NOT NULL
-        DO UPDATE SET
-            "status" = EXCLUDED."status",
-            "profile" = EXCLUDED."profile",
-            "useCases" = EXCLUDED."useCases",
-            "updated" = now()
-        WHERE "chat_personalization"."researchToken" = $7
-           OR ("chat_personalization"."status" <> ALL($8::varchar[])
-               AND "chat_personalization"."useCases" IS NULL)
-        RETURNING "id"
-        `,
-        [apId(), platformId, userId, ChatPersonalizationStatus.READY, profile, useCases, researchToken, IN_FLIGHT_STATUSES],
-    )
-    if (seeded.length === 0) {
-        log.info({ platform: { id: platformId }, user: { id: userId }, researchToken }, '[chatPersonalization] Founding-user seed skipped, the run was superseded or that row has its own research')
-    }
-}
-
-async function recoverIfStale({ row, platformId, userId, scope, log }: {
-    row: ChatPersonalization | null
-    platformId: string
-    userId: string
-    scope: ChatPersonalizationScope
-    log: FastifyBaseLogger
-}): Promise<ChatPersonalization | null> {
-    if (isNil(row)) {
-        return row
-    }
-    const inFlight = IN_FLIGHT_STATUSES.includes(row.status)
-    const stale = Date.now() - new Date(row.updated).getTime() > RESEARCH_STALENESS_TIMEOUT_MS
-    if (!inFlight || !stale) {
-        return row
-    }
-    log.warn({ platform: { id: platformId }, user: { id: userId }, scope, stuckStatus: row.status }, '[chatPersonalization] Recovering stale in-flight research row')
-    const researchToken = apId()
-    const { error } = await tryCatch(async () => {
-        const claimed = await takeOverStaleRow({ observed: row, researchToken })
-        if (isNil(claimed)) {
-            log.info({ platform: { id: platformId }, user: { id: userId }, scope }, '[chatPersonalization] Recovery abandoned, the row moved on while it was being read')
-            return
-        }
-        const allowed = await guardsAllowResearch({ platformId, log })
-        if (!allowed) {
-            await personalizationRepo().update({ id: claimed.id, researchToken }, { status: ChatPersonalizationStatus.FAILED })
-            return
-        }
-        await enqueueResearchJob({
-            platformId,
-            userId,
-            scope,
-            website: scope === ChatPersonalizationScope.COMPANY ? claimed.domain ?? null : null,
-            companyText: scope === ChatPersonalizationScope.COMPANY ? claimed.companyText ?? null : null,
-            role: claimed.role ?? null,
-            researchToken,
-            log,
-        })
-    })
-    if (error) {
-        log.warn({ error, platform: { id: platformId } }, '[chatPersonalization] Stale-row recovery failed')
-        return row
-    }
-    return findRow({ platformId, userId: scope === ChatPersonalizationScope.COMPANY ? null : userId })
-}
-
-async function takeOverStaleRow({ observed, researchToken }: {
-    observed: ChatPersonalization
-    researchToken: string
-}): Promise<ChatPersonalization | null> {
-    const swapped = await personalizationRepo()
-        .createQueryBuilder()
-        .update()
-        .set({ status: ChatPersonalizationStatus.PENDING, researchToken })
-        .where('"id" = :id', { id: observed.id })
-        .andWhere('"status" IN (:...inFlight)', { inFlight: IN_FLIGHT_STATUSES })
-        .andWhere('"updated" < now() - (:staleMs || \' milliseconds\')::interval', { staleMs: RESEARCH_STALENESS_TIMEOUT_MS })
-        .returning('*')
-        .execute()
-    return swapped.raw?.[0] ?? null
-}
-
-function normalizeRoleTitle({ input }: { input: string }): string | null {
-    const trimmed = input.trim().replace(/\s+/g, ' ')
-    if (trimmed.length === 0) {
-        return null
-    }
-    return trimmed
-        .split(' ')
-        .map((word, index) => {
-            const lower = word.toLowerCase()
-            if (ROLE_ACRONYMS.has(lower)) {
-                return lower.toUpperCase()
-            }
-            if (index > 0 && ROLE_CONNECTORS.has(lower)) {
-                return lower
-            }
-            return word.charAt(0).toUpperCase() + word.slice(1)
-        })
-        .join(' ')
-}
-
-function normalizeWebsite({ input }: { input: string }): string | null {
-    let value = input.trim().toLowerCase()
-    if (value.length === 0) {
-        return null
-    }
-    value = value.replace(/^[a-z][a-z0-9+.-]*:\/\//, '').replace(/^\/\//, '')
-    const cutAt = value.search(/[/?#:]/)
-    if (cutAt >= 0) {
-        value = value.slice(0, cutAt)
-    }
-    value = value.replace(/\.$/, '')
-    const labels = value.split('.')
-    if (labels[0] === 'www' || labels[0] === 'mail') {
-        labels.shift()
-    }
-    value = labels.join('.')
-    if (!HOSTNAME_PATTERN.test(value)) {
-        return null
-    }
-    if (/^\d+\.\d+\.\d+\.\d+$/.test(value) || value === 'localhost' || value.endsWith('.localhost') || value.endsWith('.local')) {
-        return null
-    }
-    return value
-}
-
-async function writeCompanyRow({ platformId, existing, patch }: {
-    platformId: string
-    existing: ChatPersonalization | null
-    patch: Partial<Pick<ChatPersonalization, 'domain' | 'companyText' | 'role' | 'status' | 'researchToken' | 'profile' | 'useCases'>>
-}): Promise<void> {
-    if (isNil(existing)) {
-        const { error } = await tryCatch(() => personalizationRepo().insert({
-            id: apId(),
-            platformId,
-            userId: null,
-            domain: patch.domain ?? null,
-            companyText: patch.companyText ?? null,
-            role: patch.role ?? null,
-            status: patch.status ?? ChatPersonalizationStatus.PENDING,
-            researchToken: patch.researchToken ?? null,
-            profile: patch.profile ?? null,
-            useCases: patch.useCases ?? null,
-        }))
-        if (isNil(error)) {
-            return
-        }
-    }
-    await personalizationRepo().update({ platformId, userId: IsNull() }, patch)
-}
-
-async function callerMayEditCompany({ platformId, userId, log }: {
-    platformId: string
-    userId: string
-    log: FastifyBaseLogger
-}): Promise<boolean> {
-    const user = await tryCatch(() => userService(log).getMetaInformation({ id: userId }))
-    if (user.error) {
-        log.warn({ platform: { id: platformId }, user: { id: userId }, error: user.error }, '[chatPersonalization] Could not read the platform role, leaving the company as it is')
-        return false
-    }
-    if (user.data.platformRole === PlatformRole.ADMIN) {
-        return true
-    }
-    log.info({ platform: { id: platformId }, user: { id: userId } }, '[chatPersonalization] Company edit ignored, the platform company is admin-owned')
-    return false
-}
-
-async function writeUserRow({ platformId, userId, patch }: {
-    platformId: string
-    userId: string
-    patch: Partial<Pick<ChatPersonalization, 'domain' | 'companyText' | 'role' | 'status' | 'researchToken' | 'profile' | 'useCases'>>
-}): Promise<void> {
-    const existing = await findRow({ platformId, userId })
-    if (isNil(existing)) {
-        const { error } = await tryCatch(() => personalizationRepo().insert({
-            id: apId(),
-            platformId,
-            userId,
-            domain: patch.domain ?? null,
-            companyText: patch.companyText ?? null,
-            role: patch.role ?? null,
-            status: patch.status ?? ChatPersonalizationStatus.PENDING,
-            researchToken: patch.researchToken ?? null,
-            profile: patch.profile ?? null,
-            useCases: patch.useCases ?? null,
-        }))
-        if (isNil(error)) {
-            return
-        }
-    }
-    await personalizationRepo().update({ platformId, userId }, patch)
-}
-
-async function startPrefillLookup({ platformId, userId, log }: {
-    platformId: string
-    userId: string
-    log: FastifyBaseLogger
-}): Promise<void> {
-    const { error } = await tryCatch(async () => {
-        if (isNil(apolloApiKey())) {
-            return
-        }
-        const redis = await redisConnections.useExisting()
-        const claimed = await redis.set(prefillLookupKey({ platformId, userId }), '1', 'EX', PREFILL_TTL_SECONDS, 'NX')
-        if (claimed !== 'OK') {
-            return
-        }
-        await enqueueResearchJob({
-            platformId,
-            userId,
-            scope: ChatPersonalizationScope.COMPANY,
-            website: null,
-            companyText: null,
-            role: null,
-            researchToken: null,
-            prefillOnly: true,
-            log,
-        })
-        log.info({ platform: { id: platformId }, user: { id: userId } }, '[chatPersonalization] Prefill lookup enqueued')
-    })
-    if (error) {
-        log.warn({ platform: { id: platformId }, user: { id: userId }, error }, '[chatPersonalization] Prefill lookup failed')
-    }
-}
-
-async function readPrefill({ platformId, userId, log }: {
-    platformId: string
-    userId: string
-    log: FastifyBaseLogger
-}): Promise<PersonalizationPrefill | null> {
-    const { data, error } = await tryCatch(async () => {
-        const redis = await redisConnections.useExisting()
-        const raw = await redis.get(prefillKey({ platformId, userId }))
-        if (isNil(raw)) {
-            return null
-        }
-        const parsed = PersonalizationPrefill.safeParse(JSON.parse(raw))
-        return parsed.success ? parsed.data : null
-    })
-    if (error) {
-        log.warn({ platform: { id: platformId }, user: { id: userId }, error }, '[chatPersonalization] Prefill read failed')
-        return null
-    }
-    return data
-}
-
-async function claimForResearch({ platformId, userId, scope, researchToken }: { platformId: string, userId: string, scope: PersonalizationScope, researchToken: string | null }): Promise<boolean> {
-    const scoped = scope === ChatPersonalizationScope.USER
-        ? { platformId, userId }
-        : { platformId, userId: IsNull() }
-    const criteria = {
-        ...scoped,
-        status: ChatPersonalizationStatus.PENDING,
-        ...(isNil(researchToken) ? {} : { researchToken }),
-    }
-    const updated = await personalizationRepo()
-        .createQueryBuilder()
-        .update()
-        .set({ status: ChatPersonalizationStatus.RESEARCHING })
-        .where(criteria)
-        .returning('id')
-        .execute()
-    return (updated.raw?.length ?? 0) > 0
-}
-
-async function guardsAllowResearch({ platformId, log }: { platformId: string, log: FastifyBaseLogger }): Promise<boolean> {
-    const chatProvider = await tryCatch(() => aiProviderService(log).getChatProvider({ platformId, scope: PERSONALIZATION_PROVIDER_SCOPE }))
-    if (chatProvider.error) {
-        log.warn({ platform: { id: platformId }, error: chatProvider.error }, '[chatPersonalization] Chat AI provider failed to load, skipping research')
-        return false
-    }
-    if (isNil(chatProvider.data)) {
-        log.warn({ platform: { id: platformId } }, '[chatPersonalization] No chat AI provider configured, skipping research')
-        return false
-    }
-    const credits = await tryCatch(() => assertCreditsAndAppSumoNotExceeded({ platformId, log }))
-    if (credits.error) {
-        const exhausted = credits.error instanceof ActivepiecesError && credits.error.error.code === ErrorCode.QUOTA_EXCEEDED
-        if (!exhausted) {
-            log.warn({ platform: { id: platformId }, error: credits.error }, '[chatPersonalization] Credits check failed, allowing research')
-        }
-        else {
-            log.warn({ platform: { id: platformId } }, '[chatPersonalization] Credits exhausted, skipping research')
-            return false
-        }
-    }
-    const { allowed, count } = await agentHelpers.incrementAndCheckLimit({
-        key: `chat-personalization-runs:${platformId}`,
-        limit: RESEARCH_RUNS_PER_PLATFORM_PER_DAY,
-        ttlSeconds: RATE_LIMIT_TTL_SECONDS,
-    })
-    if (!allowed) {
-        log.warn({ platform: { id: platformId }, runCount: count }, '[chatPersonalization] Daily research cap reached, skipping')
-        return false
-    }
-    return true
-}
-
-async function enqueueResearchJob({ platformId, userId, scope, website, companyText, role, researchToken, prefillOnly, log }: {
-    platformId: string
-    userId: string
-    scope: PersonalizationScope
-    website: string | null
-    companyText: string | null
-    role: string | null
-    researchToken: string | null
-    prefillOnly?: boolean
-    log: FastifyBaseLogger
-}): Promise<void> {
-    await jobQueue(log).add({
-        id: apId(),
-        type: JobType.ONE_TIME,
-        data: {
-            schemaVersion: LATEST_JOB_DATA_SCHEMA_VERSION,
-            jobType: WorkerJobType.EXECUTE_PERSONALIZATION_RESEARCH,
-            platformId,
-            projectId: null,
-            userId,
-            scope,
-            website,
-            companyText,
-            role,
-            researchToken,
-            prefillOnly: prefillOnly ?? false,
-        },
-    })
-}
-
-function prefillKey({ platformId, userId }: { platformId: string, userId: string }): string {
-    return `chat-personalization-prefill:${platformId}:${userId}`
-}
-
-function prefillLookupKey({ platformId, userId }: { platformId: string, userId: string }): string {
-    return `chat-personalization-prefill-lookup:${platformId}:${userId}`
-}
-
-function validateResult({ input, log }: { input: SavePersonalizationResultRequest, log: FastifyBaseLogger }): ValidatedResult {
-    if (input.status !== 'READY') {
-        return { status: ChatPersonalizationStatus.FAILED, profile: null, useCases: null }
-    }
-    const profile = PersonalizationProfile.safeParse(input.profile)
-    const useCases = z.array(PersonalizationUseCase).min(1).safeParse(input.useCases)
-    if (!profile.success || !useCases.success) {
-        log.warn({
-            platform: { id: input.platformId },
-            profileValid: profile.success,
-            useCasesValid: useCases.success,
-        }, '[chatPersonalization] Research result failed validation, downgrading to FAILED')
-        return { status: ChatPersonalizationStatus.FAILED, profile: null, useCases: null }
-    }
-    return { status: ChatPersonalizationStatus.READY, profile: profile.data, useCases: useCases.data }
-}
-
-function toView({ row, scope, inputsRow, role, personalStatus, prefill }: {
-    row: ChatPersonalization
-    scope: ChatPersonalizationScope
-    inputsRow: ChatPersonalization
-    role: string | null
-    personalStatus: ChatPersonalizationStatus
-    prefill: PersonalizationPrefill | null
-}): ChatPersonalizationView {
-    return {
-        status: row.status,
-        personalStatus,
-        scope,
-        useCases: row.useCases ?? [],
-        profile: row.profile ?? null,
-        companyInput: inputsRow.companyText ?? inputsRow.domain ?? null,
-        roleInput: role,
-        prefill,
-    }
-}
-
-function toScopeEnum(scope: PersonalizationScope): ChatPersonalizationScope {
-    return scope === 'user' ? ChatPersonalizationScope.USER : ChatPersonalizationScope.COMPANY
-}
-
-function emitProgress({ userId, event }: { userId: string, event: ChatPersonalizationProgressEvent }): void {
-    websocketService.to(userId).emit(WebsocketClientEvent.CHAT_PERSONALIZATION_PROGRESS, event)
-}
-
-const HOSTNAME_PATTERN = /^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)+$/
-
-const ROLE_ACRONYMS: ReadonlySet<string> = new Set(['ceo', 'cto', 'coo', 'cfo', 'cmo', 'cpo', 'cro', 'ciso', 'cio', 'chro', 'vp', 'svp', 'evp', 'hr', 'it', 'qa', 'pr', 'seo', 'sem', 'ux', 'ui', 'ai', 'ml', 'bi', 'l&d', 'r&d', 'gm', 'pm', 'gtm', 'sdr', 'bdr', 'ae', 'sre', 'csm', 'crm', 'saas', 'api'])
-
-const ROLE_CONNECTORS: ReadonlySet<string> = new Set(['of', 'and', 'the', 'for', 'in', 'at', 'to', 'a', 'an', '&'])
-
-type UpsertParams = {
-    platformId: string
-    userId: string
-    website?: string
-    role?: string
-    personalize: boolean
-}
-
-type ValidatedResult = {
-    status: ChatPersonalizationStatus
-    profile: PersonalizationProfile | null
-    useCases: PersonalizationUseCase[] | null
-}
-
-export type PersonalizationIdentity = {
-    company: PersonalizationIdentityCompany | null
-    role: string | null
-}
-
-export type PersonalizationIdentityCompany = {
-    name: string
-    description: string
-    industry: string
-}
